dev(ui): fix the directory name for <PageHeading />
