+++ src/ui/index.ts
@@ -1,2 +1,2 @@
 // Atoms
-export { default as PageHeading } from "./atoms/PageHeader/PageHeading";
+export { default as PageHeading } from "./atoms/PageHeading/PageHeading";
